requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
langgraph
langchain
anthropic
pillow
python-multipart
python-dotenv
groq